Add best and worst team season
