+++ lib/team_stats.rb
@@ -1,6 +1,8 @@
 require_relative "stats"
+require './lib/calculable'
 
 class TeamStats < Stats
+include Calculable
 
   def initialize(file_path)
     Stats.from_csv(file_path)
@@ -17,5 +19,29 @@ class TeamStats < Stats
     team_data = team.instance_variables.map { |key,value| ["#{key}".delete("@"), value = team.send("#{key}".delete("@").to_sym)]}.to_h
     # searches the hash for a key, value pair whose key is "stadium" then deletes it.
     team_data.delete_if {|k,v| k == "stadium"}
+ end
+
+ def best_season(team_id)
+   team_seasons = @@games.find_all {|game| game.home_team_id  == team_id || game.away_team_id == team_id }
+   seasons = team_seasons.group_by {|game| game.season}
+   seasons.each do |season, season_games|
+     season_game_ids = season_games.map{|game| game.game_id}
+     team_games = @@game_teams.find_all { |game| game.team_id == team_id && season_game_ids.include?(game.game_id)}
+     seasons[season] = calculate_win_percentage(team_games)
+   end
+   seasons.max_by { |season, win_pct|
+     win_pct }[0]
+ end
+
+ def worst_season(team_id)
+   team_seasons = @games.find_all {|game| game.home_team_id  == team_id || game.away_team_id == team_id }
+   seasons = team_seasons.group_by {|game| game.season}
+   seasons.each do |season, season_games|
+     season_game_ids = season_games.map{|game| game.game_id}
+     team_games = @game_teams.find_all { |game| game.team_id == team_id && season_game_ids.include?(game.game_id)}
+     seasons[season] = calculate_win_percentage(team_games)
+   end
+   seasons.min_by { |season, win_pct|
+     win_pct }[0]
  end
 end
